light refactoring of package structure
