+++ nipoppy/config/__init__.py
@@ -1,5 +1,6 @@
 """User/pipeline configurations."""
 
-from .base import Config, PipelineConfig, load_config
+from .base import Config
 from .boutiques import BoutiquesConfig
+from .pipeline import PipelineConfig
 from .singularity import SingularityConfig
